FIll order specify gas limit.
When using the test snapshot, the first exchange costs more gas. This isn't accounted for in the gas estimation so it can fail running out of gas. Set the gas limit here temporarily (assuming users will first run fill order), better fix is to run through the scenarios in the snapshot to initialize all of the storage